+++ R/globals.R
@@ -1,5 +1,13 @@
 
-utils::globalVariables(c("%>%", "dataset_id", "id", "age_group", "age_min",
-                         "age_max", "sex", "country", "state", "county", "city",
-                         "collection_start_date", "collection_end_date",
-                         "test_id", "result", "result_cat"))
+utils::globalVariables(
+  c("%>%", "dataset_id", "id", "age_group", "age_min", "age_max", "sex",
+    "country", "state", "county", "city", "collection_start_date",
+    "collection_end_date", "test_id", "result", "result_cat",
+    ":=", "NAME_0", "NAME_1", "NAME_2", "ab_denominator", "ab_titer_max",
+    "ab_titer_mean", "ab_titer_min", "ab_titer_sd", "adm1", "adm2", "age",
+    "assays_df", "denominator", "end_date", "lower_age", "numerator",
+    "pathogen", "pop_adj", "regions_df", "result_cat_pos", "seroprev",
+    "seroprev_95_ci_lower", "seroprev_95_ci_upper", "shapeID_v5", "shapeType",
+    "start_date", "test_adj", "test_combination", "test_id_1", "test_id_2",
+    "test_id_3", "upper_age")
+)