+++ inst/testing/testing.R
@@ -13,6 +13,8 @@ load_all()
 # TO DO -------------------------------------------------------------------
 
 # TODO Add github logo and url to pkgdown
+# TODO On `Reference` page of pkgdown website, put objects in meaningful groups
+# TODO Remove `clean()`
 # TODO Add source link to each vignette
 # TODO Explore creating a specific class for the output
 # TODO Consider using readr::parse_number() for `age` and `result` values
@@ -59,7 +61,7 @@ estimates <- st_aggregate(validated_df, add_ci = TRUE)
 
 
 
-# 95% CI ------------------------------------------------------------------
+# 95% CI - bootstrap ------------------------------------------------------
 
 dat_ci <- dat %>%
   tidyr::drop_na(result_cat) %>%
@@ -101,35 +103,57 @@ b
 
 
 
-# Binomial 95% CI ---------------------------------------------------------
+# st_locate() -------------------------------------------------------------
 
-validated_df <- validated_df %>%
-  tidyr::drop_na(result) %>%
-  dplyr::mutate(
-    lower = as.data.frame(Hmisc::binconf(4, 22))$Lower,
-    upper = as.data.frame(Hmisc::binconf(4, 22))$Upper
-  )
+st_locate(sample_raw_data, adm0 = regions$adm0$Canada,
+          adm1 = adm1, #adm2 = adm2,
+          "bc" = regions$adm1$Canada$`British Columbia`,
+          "Chernihiv Oblast" = regions$adm1$Canada$`British Columbia`,
+          "Tarxien" = regions$adm1$Canada$`British Columbia`,
+          "Chari-Baguirmi" = regions$adm1$Canada$`British Columbia`,
+          "Blida" = regions$adm1$Canada$`British Columbia`
+          )
 
-a <- Hmisc::binconf(22, 45) %>% as.data.frame()
-
-as.data.frame(Hmisc::binconf(22, 45))$Lower
-as.data.frame(Hmisc::binconf(22, 45))$Upper
-
-num <- c(22, 33, 44)
-denom <- c(45, 55, 65)
-binom.test(num, denom)
-Hmisc::binconf(num, denom)
-binom::binom.confint(num, denom, methods = "wilson")$lower
-binom::binom.confint(num, denom, methods = "wilson")$upper
-
-
-p <- num/denom
-p + c(-qnorm(0.975),qnorm(0.975))*sqrt((1/denom)*p*(1-p))
+st_locate(sample_raw_data, adm0 = regions$adm0$Canada,
+          adm1 = regions$adm1$Canada$Alberta, adm2 = adm2,
+          # "Isles of Scilly" = regions$adm2$Canada$Manitoba$Winnipeg,
+          # "Malheur" = regions$adm2$Canada$Manitoba$Winnipeg,
+          # "Go Cong Tay" = regions$adm2$Canada$Manitoba$Winnipeg,
+          # "Reduit Park" = regions$adm2$Canada$Manitoba$Winnipeg,
+          # "Rondón" = regions$adm2$Canada$Manitoba$Winnipeg,
+          # "Um Bada" = regions$adm2$Canada$Manitoba$Winnipeg,
+          # "Kasan" = regions$adm2$Canada$Manitoba$Winnipeg,
+          into = c("adm1_new", "adm2_new"),
+          n_unmatched_printed = 3.00001)
 
 
-b <- binom.test(22, 45)
-attributes(b)
-b$conf.int[1]
-b$conf.int[2]
+st_locate(sample_raw_data, adm0 = regions$adm0$Canada,
+          adm1 = regions$adm1$Canada$Alberta,
+          adm2 = regions$adm2$Canada$Alberta$Calgary,
+          "ab" = regions$adm1$Canada$Yukon)
 
-22/45
+
+
+zzz <- st_locate(sample_raw_data, adm0 = regions$adm0$Canada,
+                 adm1 = adm1, adm2 = adm2)
+mydata <- tibble(
+  adm1 = c("alberta", "california", "bc"),
+  adm2 = c(NA, NA, "Vancouver"),
+  adm1_code = NA_character_, adm2_code = NA_character_
+)
+
+
+country <- "Canada"
+lhs <- c("alberta", "Vancouver")
+rhs <- c(regions$adm1$Canada$Alberta,
+         regions$adm2$Canada$`British Columbia`$`Vancouver Island and Coast / *`)
+
+evaluate_lhs_rhs(data = mydata, lhs = lhs, rhs = rhs, adm0_code = regions$adm0$Canada,
+                 adm1 = "state", adm2 = "district")
+
+st_locate(mydata, adm0 = regions$adm0$Canada, adm1 = adm1, adm2 = adm2,
+          "california" = regions$adm1$Canada$Alberta,
+          "bc" = regions$adm1$Canada$`British Columbia`,
+          "Vancouver" = regions$adm2$Canada$`British Columbia`$Nechako,
+          into = c("adm1_code", "adm2_code"))
+