+++ NEWS.md
@@ -1,5 +1,9 @@
 # serotrackr (development version)
 
+# serotrackr 0.2.0
+
+* Added `st_locate()` to help with converting region names to `serotrackr`'s acceptable region codes. First, the function has an automatic component where it tries to match the user's region names to those predefined in `serotrackr::regions_df`. This is an exact case insensitive match that treats accented and unaccented letters the same. The function then produces a report of the user's unmatched unique region names. In its second component, the function has a dot-dot-dot (ellipsis) argument where region codes can be defined for each unmatched unique region name one by one.
+* Added an argument to `st_aggregate()` to generate 95% binomial proportion confidence interval for the seroprevalence estimate.
 * `st_save()` replaced `save_xlsx()`. Apart from a more consistent naming scheme, the new function uses the output of both `st_validate()` and `st_aggregate()` functions to fill in as much cells as possible in the Excel template.
 * Added initial version of `st_aggregate()` to generate aggregate estimates from validated individual level data.
 * `st_validate()` replaced `map_cols()`. Naming of functions will be more consistent, while trying to avoid namespace conflicts. The new function accepts both columns and single values, checks for required input types (character, numeric, or date), applies several validation `rules` for each argument, and generates descriptive messages in the console about what parts of data are okay and what parts have issues.